Final Case Study-OnlineShoppingCart
A shopping cart system is a virtual store on the Internet where customers can login, browse the products and select products of interest. The selected items will be moved to cart. At checkout time, the items in the cart will be presented as an order. To complete the transaction the customer will be asked to fill a shipping address, and payment information